+++ CHANGELOG.md
@@ -7,10 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.9.0] - 2025-08-22
+
 ### Added
 
 - `Capabilities`: Added `hasConformance` function and `Conformance` URI object
 - `Connection`: Added `listProcessingParameters` function
+
+### Fixed
+
+- Typescript declaration
 
 ## [2.8.0] - 2025-04-25
 
@@ -254,7 +260,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 All prior releases have been documented in the [GitHub Releases](https://github.com/Open-EO/openeo-js-client/releases).
 
-[Unreleased]: https://github.com/Open-EO/openeo-js-client/compare/v2.8.0...HEAD
+[Unreleased]: https://github.com/Open-EO/openeo-js-client/compare/v2.9.0...HEAD
+[2.9.0]: https://github.com/Open-EO/openeo-js-client/compare/v2.8.0...v2.9.0
 [2.8.0]: https://github.com/Open-EO/openeo-js-client/compare/v2.7.0...v2.8.0
 [2.7.0]: https://github.com/Open-EO/openeo-js-client/compare/v2.6.0...v2.7.0
 [2.6.0]: https://github.com/Open-EO/openeo-js-client/compare/v2.5.1...v2.6.0